Move lockdown tests to separate file, also test background
